requirements.txt
yfinance>=0.2.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
pyyaml>=6.0
scipy>=1.9.0
tabulate>=0.9.0
